Changed booking validation error message in cart
The mechanism that extracts the messages from the validation exception has not been removed.
A new method has been added as extra abstraction.
The handler uses this method to retrieve the errors to register with EDD, regardless of origin.
That origin has been changed in this commit to be a single generic error message.